Update base.html
I saw a few issues in your HTML code. I've point them out and provided corrections:

The link tag for the Bootstrap stylesheet has a typo. It should be <link> instead of <lnk>.

The script tag for your JavaScript file has a typo. It should be <script> instead of <scrpt>.

You have a closing </htl> tag, but it should be </html> to properly close the HTML document.

You have a closing </hea> tag, but it should be </head>